Moved all variant related files into own directory called variants
