+++ public/src/components/channelManagement/variants/testEditorVariantSummaryWebPreviewButton.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Button } from '@mui/material';
 import VisibilityIcon from '@mui/icons-material/Visibility';
-import { TestPlatform, TestType } from './helpers/shared';
-import { getStage, Stage } from '../../utils/stage';
+import { TestPlatform, TestType } from '../helpers/shared';
+import { getStage, Stage } from '../../../utils/stage';
 
 export type ArticleType = 'Standard' | 'Liveblog';
